+++ packages/events/src/index.ts
@@ -12,6 +12,7 @@ export type {
   Fetch,
   ProgramsResource,
   RequestOptions,
+  RetryOptions,
   SeasonsResource,
   TeamsResource,
   VexEventsClientOptions,